+++ dflib-benchmark/src/main/java/org/dflib/benchmark/speed/DataFrameMergeInt.java
@@ -1,7 +1,7 @@
 package org.dflib.benchmark.speed;
 
 import org.dflib.DataFrame;
-import org.dflib.Series;
+import org.dflib.IntSeries;
 import org.dflib.benchmark.ValueMaker;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
@@ -23,7 +23,7 @@ import java.util.concurrent.TimeUnit;
 @OutputTimeUnit(TimeUnit.MILLISECONDS)
 @Fork(2)
 @State(Scope.Thread)
-public class DataFrameMapInteger {
+public class DataFrameMergeInt {
 
     @Param("1000000")
     public int rows;
@@ -32,7 +32,7 @@ public class DataFrameMapInteger {
 
     @Setup
     public void setUp() {
-        Series<Integer> c0 = ValueMaker.randomIntSeq((rows) / 2).series(rows);
+        IntSeries c0 = ValueMaker.intSeq().intSeries(rows);
         df = DataFrame.byColumn("c0").of(c0);
     }
 
@@ -40,7 +40,7 @@ public class DataFrameMapInteger {
     public Object mapCast() {
         return df
                 .cols()
-                .map((r, t) -> t.set(0, ((int) r.get(0)) * 10))
+                .merge((r, t) -> t.set(0, ((int) r.get(0)) * 10))
                 .materialize()
                 .iterator();
     }
@@ -49,7 +49,17 @@ public class DataFrameMapInteger {
     public Object mapApiCast() {
         return df
                 .cols()
-                .map((r, t) -> t.set(0, r.get(0, Integer.class) * 10))
+                .merge((r, t) -> t.set(0, r.get(0, Integer.class) * 10))
+                .materialize()
+                .iterator();
+    }
+
+
+    @Benchmark
+    public Object mapInt() {
+        return df
+                .cols()
+                .merge((r, t) -> t.set(0, r.getInt(0) * 10))
                 .materialize()
                 .iterator();
     }