+++ dflib-jdbc/src/main/java/org/dflib/jdbc/connector/saver/SaveViaUpsert.java
@@ -120,8 +120,8 @@ public class SaveViaUpsert extends TableSaveStrategy {
         // TODO: speed up the equality test by excluding "keyColumns" from both sides
 
         // note that "toSave" and "previouslySaved" must be ordered by key for "eq" to be meaningful
-        DataFrame eqMatrix = toSave.eq(previouslySaved).colsAppend(DIFF_COLUMN).map(this::booleansAsBitSet);
-        DataFrame toSaveClassified = toSave.colsAppend(DIFF_COLUMN).map(eqMatrix.getColumn(DIFF_COLUMN));
+        DataFrame eqMatrix = toSave.eq(previouslySaved).colsAppend(DIFF_COLUMN).merge(this::booleansAsBitSet);
+        DataFrame toSaveClassified = toSave.colsAppend(DIFF_COLUMN).merge(eqMatrix.getColumn(DIFF_COLUMN));
 
         infoTracker.updatesCardinality(toSaveClassified.getColumn(DIFF_COLUMN));