+++ dflib/src/test/java/org/dflib/ColumnSet_Merge_RowToValueTest.java
@@ -3,29 +3,28 @@ package org.dflib;
 import org.dflib.unit.DataFrameAsserts;
 import org.junit.jupiter.api.Test;
 
-import static org.dflib.Exp.*;
+public class ColumnSet_Merge_RowToValueTest {
 
-public class ColumnSet_Map_ExpTest {
+    @Test
+    public void colsAppend() {
+        DataFrame df = DataFrame.foldByRow("a", "b")
+                .of(1, "x", 2, "y")
+                .colsAppend("b", "c")
+                .merge(r -> r.get(0, Integer.class) * 100, r -> r.get(0, Integer.class) * 10);
+
+        new DataFrameAsserts(df, "a", "b", "b_", "c")
+                .expectHeight(2)
+                .expectRow(0, 1, "x", 100, 10)
+                .expectRow(1, 2, "y", 200, 20);
+    }
 
     @Test
     public void cols_ByName() {
         DataFrame df = DataFrame.foldByRow("a", "b")
                 .of(1, "x", 2, "y")
-                .cols("b", "c").map($int(0).mul(100), $int(0).mul(10));
+                .cols("b", "c").merge(r -> r.get(0, Integer.class) * 100, r -> r.get(0, Integer.class) * 10);
 
         new DataFrameAsserts(df, "a", "b", "c")
-                .expectHeight(2)
-                .expectRow(0, 1, 100, 10)
-                .expectRow(1, 2, 200, 20);
-    }
-
-    @Test
-    public void cols_ByName_Duplicate() {
-        DataFrame df = DataFrame.foldByRow("a", "b")
-                .of(1, "x", 2, "y")
-                .cols("b", "b").map($int(0).mul(100), $int(0).mul(10));
-
-        new DataFrameAsserts(df, "a", "b", "b_")
                 .expectHeight(2)
                 .expectRow(0, 1, 100, 10)
                 .expectRow(1, 2, 200, 20);
@@ -35,7 +34,7 @@ public class ColumnSet_Map_ExpTest {
     public void cols_ByPos() {
         DataFrame df = DataFrame.foldByRow("a", "b")
                 .of(1, "x", 2, "y")
-                .cols(1, 2).map($int(0).mul(100), $int(0).mul(10));
+                .cols(1, 2).merge(r -> r.get(0, Integer.class) * 100, r -> r.get(0, Integer.class) * 10);
 
         new DataFrameAsserts(df, "a", "b", "2")
                 .expectHeight(2)
@@ -44,37 +43,22 @@ public class ColumnSet_Map_ExpTest {
     }
 
     @Test
-    public void colsAppend() {
-        DataFrame df = DataFrame.foldByRow("a", "b")
-                .of(1, "x", 2, "y")
-                .colsAppend("b", "c", "b").map(
-                        $int(0).mul(100),
-                        $int(0).mul(10),
-                        $str(1).mapVal(v -> "[" + v + "]"));
-
-        new DataFrameAsserts(df, "a", "b", "b_", "c", "b__")
-                .expectHeight(2)
-                .expectRow(0, 1, "x", 100, 10, "[x]")
-                .expectRow(1, 2, "y", 200, 20, "[y]");
-    }
-
-    @Test
     public void cols() {
         DataFrame df = DataFrame.foldByRow("a", "b")
                 .of(1, "x", 2, "y")
-                .cols().map($int(0).mul(100).as("b"), $int(0).mul(10));
+                .cols().merge(r -> r.get(0, Integer.class) * 100, r -> r.get(0, Integer.class) * 10);
 
-        new DataFrameAsserts(df, "a", "b", "a * 10")
+        new DataFrameAsserts(df, "a", "b", "2", "3")
                 .expectHeight(2)
-                .expectRow(0, 1, 100, 10)
-                .expectRow(1, 2, 200, 20);
+                .expectRow(0, 1, "x", 100, 10)
+                .expectRow(1, 2, "y", 200, 20);
     }
 
     @Test
     public void colsExcept_ByName() {
         DataFrame df = DataFrame.foldByRow("a", "b", "c")
                 .of(1, "x", "z", 2, "y", "a")
-                .colsExcept("a").map($int(0).mul(100), $int(0).mul(10));
+                .colsExcept("a").merge(r -> r.get(0, Integer.class) * 100, r -> r.get(0, Integer.class) * 10);
 
         new DataFrameAsserts(df, "a", "b", "c")
                 .expectHeight(2)
@@ -86,7 +70,7 @@ public class ColumnSet_Map_ExpTest {
     public void colsExcept_ByPos() {
         DataFrame df = DataFrame.foldByRow("a", "b", "c")
                 .of(1, "x", "z", 2, "y", "a")
-                .colsExcept(1).map($int(0).mul(100), $int(0).mul(10));
+                .colsExcept(1).merge(r -> r.get(0, Integer.class) * 100, r -> r.get(0, Integer.class) * 10);
 
         new DataFrameAsserts(df, "a", "b", "c")
                 .expectHeight(2)