+++ dflib/src/test/java/org/dflib/RowSet_Merge_MappersTest.java
@@ -3,7 +3,7 @@ package org.dflib;
 import org.dflib.unit.DataFrameAsserts;
 import org.junit.jupiter.api.Test;
 
-public class RowSet_Map_RowMapperTest {
+public class RowSet_Merge_MappersTest {
 
     @Test
     public void all() {
@@ -14,11 +14,11 @@ public class RowSet_Map_RowMapperTest {
                         -1, "m", "n")
                 .cols(0).compactInt(0)
                 .rows()
-                .map((f, t) -> {
-                    t.set(0, f.getInt(0) * 3);
-                    t.set(1, f.get(1) + "" + f.get(2));
-                    t.set(2, f.get(2));
-                });
+                .merge(
+                        r -> r.getInt(0) * 3,
+                        r -> r.get(1) + "" + r.get(2),
+                        r -> r.get(2)
+                );
 
         new DataFrameAsserts(df, "a", "b", "c")
                 .expectHeight(3)
@@ -36,11 +36,11 @@ public class RowSet_Map_RowMapperTest {
                         -1, "m", "n")
                 .cols(0).compactInt(0)
                 .rows(Series.ofInt(0, 2))
-                .map((f, t) -> {
-                    t.set(0, f.getInt(0) * 3);
-                    t.set(1, f.get(1) + "" + f.get(2));
-                    t.set(2, f.get(2));
-                });
+                .merge(
+                        r -> r.getInt(0) * 3,
+                        r -> r.get(1) + "" + r.get(2),
+                        r -> r.get(2)
+                );
 
         new DataFrameAsserts(df, "a", "b", "c")
                 .expectHeight(3)
@@ -58,11 +58,11 @@ public class RowSet_Map_RowMapperTest {
                         -1, "m", "n")
                 .cols(0).compactInt(0)
                 .rows(Series.ofInt(0, 2, 2, 0))
-                .map((f, t) -> {
-                    t.set(0, f.getInt(0) * 3);
-                    t.set(1, f.get(1) + "" + f.get(2));
-                    t.set(2, f.get(2));
-                });
+                .merge(
+                        r -> r.getInt(0) * 3,
+                        r -> r.get(1) + "" + r.get(2),
+                        r -> r.get(2)
+                );
 
         new DataFrameAsserts(df, "a", "b", "c")
                 .expectHeight(5)
@@ -81,41 +81,20 @@ public class RowSet_Map_RowMapperTest {
                         2, "y", "b",
                         -1, "m", "n")
                 .cols(0).compactInt(0)
-                .rowsRange(1, 2)
-                .map((f, t) -> {
-                    t.set(0, f.getInt(0) * 3);
-                    t.set(1, f.get(1) + "" + f.get(2));
-                    t.set(2, f.get(2));
-                });
-
-        new DataFrameAsserts(df, "a", "b", "c")
-                .expectHeight(3)
-                .expectRow(0, 1, "x", "a")
-                .expectRow(1, 6, "yb", "b")
-                .expectRow(2, -1, "m", "n");
-    }
-
-    @Test
-    public void byIndex_Unordered() {
-        DataFrame df = DataFrame.foldByRow("a", "b", "c")
-                .of(
-                        1, "x", "a",
-                        2, "y", "b",
-                        -1, "m", "n")
-                .cols(0).compactInt(0)
-                .rows(Series.ofInt(2, 0))
-                .map((f, t) -> {
-                    t.set(0, f.getInt(0) * 3);
-                    t.set(1, f.get(1) + "" + f.get(2));
-                    t.set(2, f.get(2));
-                });
+                .rowsRange(0, 2)
+                .merge(
+                        r -> r.getInt(0) * 3,
+                        r -> r.get(1) + "" + r.get(2),
+                        r -> r.get(2)
+                );
 
         new DataFrameAsserts(df, "a", "b", "c")
                 .expectHeight(3)
                 .expectRow(0, 3, "xa", "a")
-                .expectRow(1, 2, "y", "b")
-                .expectRow(2, -3, "mn", "n");
+                .expectRow(1, 6, "yb", "b")
+                .expectRow(2, -1, "m", "n");
     }
+
 
     @Test
     public void byCondition() {
@@ -126,11 +105,11 @@ public class RowSet_Map_RowMapperTest {
                         -1, "m", "n")
                 .cols(0).compactInt(0)
                 .rows(Series.ofBool(true, false, true))
-                .map((f, t) -> {
-                    t.set(0, f.getInt(0) * 3);
-                    t.set(1, f.get(1) + "" + f.get(2));
-                    t.set(2, f.get(2));
-                });
+                .merge(
+                        r -> r.getInt(0) * 3,
+                        r -> r.get(1) + "" + r.get(2),
+                        r -> r.get(2)
+                );
 
         new DataFrameAsserts(df, "a", "b", "c")
                 .expectHeight(3)