+++ dflib/src/test/java/org/dflib/RowSet_Merge_RowMapperTest.java
@@ -3,7 +3,7 @@ package org.dflib;
 import org.dflib.unit.DataFrameAsserts;
 import org.junit.jupiter.api.Test;
 
-public class RowSet_Map_ExpTest {
+public class RowSet_Merge_RowMapperTest {
 
     @Test
     public void all() {
@@ -12,12 +12,13 @@ public class RowSet_Map_ExpTest {
                         1, "x", "a",
                         2, "y", "b",
                         -1, "m", "n")
+                .cols(0).compactInt(0)
                 .rows()
-                .map(
-                        Exp.$int(0).mul(3),
-                        Exp.concat(Exp.$str(1), Exp.$str(2)),
-                        Exp.$str(2)
-                );
+                .merge((f, t) -> {
+                    t.set(0, f.getInt(0) * 3);
+                    t.set(1, f.get(1) + "" + f.get(2));
+                    t.set(2, f.get(2));
+                });
 
         new DataFrameAsserts(df, "a", "b", "c")
                 .expectHeight(3)
@@ -33,39 +34,19 @@ public class RowSet_Map_ExpTest {
                         1, "x", "a",
                         2, "y", "b",
                         -1, "m", "n")
+                .cols(0).compactInt(0)
                 .rows(Series.ofInt(0, 2))
-                .map(
-                        Exp.$int(0).mul(3),
-                        Exp.concat(Exp.$str(1), Exp.$str(2)),
-                        Exp.$str(2)
-                );
+                .merge((f, t) -> {
+                    t.set(0, f.getInt(0) * 3);
+                    t.set(1, f.get(1) + "" + f.get(2));
+                    t.set(2, f.get(2));
+                });
 
         new DataFrameAsserts(df, "a", "b", "c")
                 .expectHeight(3)
                 .expectRow(0, 3, "xa", "a")
                 .expectRow(1, 2, "y", "b")
                 .expectRow(2, -3, "mn", "n");
-    }
-
-    @Test
-    public void rowsExcept_byIndex() {
-        DataFrame df = DataFrame.foldByRow("a", "b", "c")
-                .of(
-                        1, "x", "a",
-                        2, "y", "b",
-                        -1, "m", "n")
-                .rowsExcept(Series.ofInt(0, 2))
-                .map(
-                        Exp.$int(0).mul(3),
-                        Exp.concat(Exp.$str(1), Exp.$str(2)),
-                        Exp.$str(2)
-                );
-
-        new DataFrameAsserts(df, "a", "b", "c")
-                .expectHeight(3)
-                .expectRow(0, 1, "x", "a")
-                .expectRow(1, 6, "yb", "b")
-                .expectRow(2, -1, "m", "n");
     }
 
     @Test
@@ -77,11 +58,11 @@ public class RowSet_Map_ExpTest {
                         -1, "m", "n")
                 .cols(0).compactInt(0)
                 .rows(Series.ofInt(0, 2, 2, 0))
-                .map(
-                        Exp.$int(0).mul(3),
-                        Exp.concat(Exp.$str(1), Exp.$str(2)),
-                        Exp.$str(2)
-                );
+                .merge((f, t) -> {
+                    t.set(0, f.getInt(0) * 3);
+                    t.set(1, f.get(1) + "" + f.get(2));
+                    t.set(2, f.get(2));
+                });
 
         new DataFrameAsserts(df, "a", "b", "c")
                 .expectHeight(5)
@@ -99,17 +80,40 @@ public class RowSet_Map_ExpTest {
                         1, "x", "a",
                         2, "y", "b",
                         -1, "m", "n")
-                .rowsRange(1, 3)
-                .map(
-                        Exp.$int(0).mul(3),
-                        Exp.concat(Exp.$str(1), Exp.$str(2)),
-                        Exp.$str(2)
-                );
+                .cols(0).compactInt(0)
+                .rowsRange(1, 2)
+                .merge((f, t) -> {
+                    t.set(0, f.getInt(0) * 3);
+                    t.set(1, f.get(1) + "" + f.get(2));
+                    t.set(2, f.get(2));
+                });
 
         new DataFrameAsserts(df, "a", "b", "c")
                 .expectHeight(3)
                 .expectRow(0, 1, "x", "a")
                 .expectRow(1, 6, "yb", "b")
+                .expectRow(2, -1, "m", "n");
+    }
+
+    @Test
+    public void byIndex_Unordered() {
+        DataFrame df = DataFrame.foldByRow("a", "b", "c")
+                .of(
+                        1, "x", "a",
+                        2, "y", "b",
+                        -1, "m", "n")
+                .cols(0).compactInt(0)
+                .rows(Series.ofInt(2, 0))
+                .merge((f, t) -> {
+                    t.set(0, f.getInt(0) * 3);
+                    t.set(1, f.get(1) + "" + f.get(2));
+                    t.set(2, f.get(2));
+                });
+
+        new DataFrameAsserts(df, "a", "b", "c")
+                .expectHeight(3)
+                .expectRow(0, 3, "xa", "a")
+                .expectRow(1, 2, "y", "b")
                 .expectRow(2, -3, "mn", "n");
     }
 
@@ -120,12 +124,13 @@ public class RowSet_Map_ExpTest {
                         1, "x", "a",
                         2, "y", "b",
                         -1, "m", "n")
+                .cols(0).compactInt(0)
                 .rows(Series.ofBool(true, false, true))
-                .map(
-                        Exp.$int(0).mul(3),
-                        Exp.concat(Exp.$str(1), Exp.$str(2)),
-                        Exp.$str(2)
-                );
+                .merge((f, t) -> {
+                    t.set(0, f.getInt(0) * 3);
+                    t.set(1, f.get(1) + "" + f.get(2));
+                    t.set(2, f.get(2));
+                });
 
         new DataFrameAsserts(df, "a", "b", "c")
                 .expectHeight(3)